Fix container resize drag for touchscreens. (#75)
Container resize drags do not work on touchscreens because
ui.interact(...) is called AFTER the drag event started. This PR moves
the ui.interact(...) call for the "resize drag handle" out of the
current IF block so that the start of the drag event can be registered
properly.

This was mostly not a problem for mouse-based interaction because the IF
condition on 'latest_pos()' returns true most of the time as long as the
mouse pointer is hovering over the screen input area. This is of course
not the case with touchscreens, hence additionally 'latest_pos()' was
replaced by 'interact_pos()'.

Fix tested on:
wasm + touchscreen
wasm + mouse
native + mouse