Add `Tree::inactive_tiles` for getting non-visible tiles (#106)
Hi!
 
I really liked the project and decided to add a counterpart to the
`active_tiles` method to the `Tree`, since I found it useful to have it
alongside the other method. I think it's nice if this is provided by the
crate itself, since a user might mix-up checking the _visible_ tiles
using `is_visible` with a `Tile` actually being _inactive_ and therefore
also _not-visible_, so having this functionality being provided as an
analogous to the `active_tiles` method could make it easier on the user
to avoid confusion (it confused me!).

Also fixed some minor typo.

Thank you so much for this project!

---------

Co-authored-by: Emil Ernerfeldt <emil.ernerfeldt@gmail.com>